:recycle: Refactor components for reusability
