+++ interactive/src/components/Tier.jsx
@@ -1,6 +1,8 @@
 import { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { Container, List, ListItem, Tooltip, Typography } from '@mui/material';
+import { CheckboxItem } from './Checklist';
+import ProgressLabelled from './ProgressLabelled';
 
 class Tier extends PureComponent {
   static propTypes = {
@@ -8,7 +10,20 @@ class Tier extends PureComponent {
   }
 
   render() {
-    const { tier, tiers } = this.props;
+    const { checkDenominator, checkNumerator, checks, handleCheck, items, section, tier, tiers } = this.props;
+
+    const prerequisites = (tier, section) => {
+      if (!tiers[tier].hasOwnProperty('prerequisiteTiers')) {
+        return null;
+      } else {
+        return tiers[tier].prerequisiteTiers.map((prereq, index) => {
+          const denominator = checkDenominator(tier, prereq, section);
+          return (
+            <ProgressLabelled key={ tier + "-" + section + "-prerequisites-" + index.toString() } tier={ prereq } numerator={ checkNumerator(prereq, section) } denominator={ denominator } />
+          );
+        });
+      }
+    }
 
     const title = Object.keys(tiers).filter(_tier => _tier === tier).map(_tier => {
       return (
@@ -22,11 +37,24 @@ class Tier extends PureComponent {
         </Container>
       );
     });
-      
+
     return (
-      <Tooltip key={tier + "-benefits-tooltip"} title={ title }>
-        <Typography variant="h3">{ tier }</Typography>
-      </Tooltip>
+      <Container key={tier + '=' + section}>
+        <Tooltip key={tier + "-benefits"} title={ title }>
+          <Typography variant="h3">{ tier }</Typography>
+        </Tooltip>
+        <List>
+          { prerequisites(tier, section) }
+          { items.filter(item => item.section === section && item.tier === tier).map(item => (
+              <CheckboxItem
+                key={ item.prompt }
+                item={ item }
+                checks={ checks }
+                handleCheck={ handleCheck }
+              />
+          ))}
+        </List>
+      </Container>
     );
   }
 }
